Se agregan test unitarios a las capas del proyecto

+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/com/projectve/r2dbc/MyReactiveRepository.java
@@ -4,9 +4,8 @@ import co.com.projectve.r2dbc.entity.CreditApplicationEntity;
 import org.springframework.data.repository.query.ReactiveQueryByExampleExecutor;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 
-import java.math.BigInteger;
 
 // TODO: This file is just an example, you should delete or modify it
-public interface MyReactiveRepository extends ReactiveCrudRepository<CreditApplicationEntity, BigInteger>, ReactiveQueryByExampleExecutor<CreditApplicationEntity> {
+public interface MyReactiveRepository extends ReactiveCrudRepository<CreditApplicationEntity, Integer>, ReactiveQueryByExampleExecutor<CreditApplicationEntity> {
 
 }
